examples/psycopg-postgres/requirements.txt
dbop-core[psycopg]
psycopg[binary]>=3.1.8
python-dotenv
